Add feature that allows auto-typing to libvirt using virDomainSendKey.
This change extends the auto-type implementation to allow multiple
auto-type plugins to live alongside the platform-native plugins, and
the auto-type to libvirt is implemeneted with such a plugin.

The difference with the platform-native plugins is that these
'external' plugins execute the auto-type on a selected target. The
plugins provide a list of available targets, which are presented
to the user as target selection.

When guests targeted by the libvirt plugin use a keyboard layout
with dead keys, certain characters need a trailing space keycode to
properly output the character. Because Windows seems to be the only
OS that uses dead keys in the default keyboard layout (US-intl),
two new options added to the settings page. One to enable dead keys
if a Windows guest is detected (on by default), and the other to
enable dead keys for other guest operating systems (off by default).